feat: images for Cloud Hypervisor
This change includes a set of new images for use with Cloud Hypervisor.
Specifically the following types of images:

- Kernel
- Base Ubuntu
- Cluster API Provider Microvm

These have been added to a `experimental` folder as its antivcapted
these will be replaced in the future with a better image-building
mechanism (that has already been proved in a poc).

Signed-off-by: Richard Case <richard.case@outlook.com>