ref: Do not build capi image from kernel base (#52)
The kernel base image brought the kernel modules into the operating
system image. This resulted in quite a long image pipeline, as a change
in any of the kernel layers would require updating the OS layers.
Our new plan is to mount those modules in as a separate volume, thus
separating the kernel image pipeline from the OS image pipeline.

This change updates the CAPI OS image so that it is built from a
standard Ubuntu instead. It does not contain module files.

In order not to break things for people who may be using versions of
CAPMVM and Flintlock which do not have the feature to add modules via an
additional mountpoint, I have renamed the image. This will ensure that
the existing tags will not be overwritten. I chose this as the less
confusing and frustrating option for people. I am open to any other
suggestions on how to do this. Another way could be suffixing the image
tags with `-v2` or something.